Creacion de pedidos

+++ database/migrations/2023_01_29_pedidos.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
 
             $table->string("fechaPedido");
-            $table->integer("total");
+            $table->double("total");
 
             $table->rememberToken();
             $table->timestamps();
